+++ json/src/test/scala/com/paypal/cascade/json/tests/JsonUtilSpecs.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
- package com.paypal.cascade.json.tests
+package com.paypal.cascade.json.tests
 
 import scala.util.Try
 
@@ -26,10 +26,11 @@ import org.scalacheck.Prop._
 import org.specs2._
 
 import com.paypal.cascade.common.tests.scalacheck._
-import com.paypal.cascade.json.JsonUtil._
+import com.paypal.cascade.json._
 
 /**
- * Tests that exercise methods in [[com.paypal.cascade.json.JsonUtil]]
+ * Tests that exercise methods in [[com.paypal.cascade.json.JsonUtil]] by way of
+ * implicit wrapper methods in the [[com.paypal.cascade.json]] package object.
  */
 class JsonUtilSpecs
   extends Specification
@@ -56,6 +57,7 @@ class JsonUtilSpecs
 
   JsonUtil should serialize and deserialize case classes, such as
     a case class containing a single data member                             ${CaseClasses.OneMember().ok}
+    a case class containing a single polymorphic member                      ${CaseClasses.OneMemberAny().ok}
     a case class containing multiple members of mixed basic types            ${CaseClasses.TwoMemberMixedBasic().ok}
     a case class containing mutliple members of mixed complex types          ${CaseClasses.TwoMemberMixedComplex().ok}
     a case class containing an optional AnyVal type                          ${CaseClasses.OptionalAnyValMember().ok}
@@ -67,124 +69,128 @@ class JsonUtilSpecs
   JsonUtil should
     not deserialize malformed json                                           ${Badness.MalformedJson().fails}
     not deserialize json that is type mismatched                             ${Badness.MismatchedTypes().fails}
+    not convert between incompatible types                                   ${Badness.BadValueConversions().fails}
+    deserialize null from Any with None, and deserialize Some(_) otherwise   ${Badness.AnyToNullConversion().ok}
     deserialize json that is missing an AnyVal, with a default value         ${Badness.MissingAnyVal().ok}
     deserialize json that is missing an AnyRef, with a null value            ${Badness.MissingAnyRef().ok}
-
-  Implicit class tests
-
-  toJson and fromJson methods work directly on strings                       ${ImplicitClasses().success}
 
   """
 
   object BasicTypes {
 
+    private def basicMatcher[T : Manifest](in: T, expectedJson: String) = {
+      val to = in.toJson.get
+      val from = to.fromJson[T].get
+      val toConvert: Any = from.asInstanceOf[Any]
+      val converted = toConvert.convertValue[T].get
+      (to must beEqualTo(expectedJson)) and
+        (from must beEqualTo(in)) and
+        (converted must beEqualTo(from))
+    }
+
     case class Strings() {
       def ok = forAll(genJsonString) { str =>
-        val to = toJson(str).get
-        val from = fromJson[String](to).get
-        // for whatever reason, the compiler balks on string interpolation here
-        // e.g. "\"$str\"" is not recognized as a valid string
-        (to must beEqualTo("\"%s\"".format(str))) and (from must beEqualTo(str))
-      }
+        basicMatcher(str, "\"%s\"".format(str)) }
     }
 
     case class Ints() {
       def ok = forAll(arbitrary[Int]) { i =>
-        val to = toJson(i).get
-        val from = fromJson[Int](to).get
-        (to must beEqualTo(i.toString)) and (from must beEqualTo(i))
-      }
+        basicMatcher(i, i.toString) }
     }
 
     case class Longs() {
       def ok = forAll(arbitrary[Long]) { i =>
-        val to = toJson(i).get
-        val from = fromJson[Long](to).get
-        (to must beEqualTo(i.toString)) and (from must beEqualTo(i))
-      }
+        basicMatcher(i, i.toString) }
     }
 
     case class Floats() {
       def ok = forAll(arbitrary[Float]) { i =>
-        val to = toJson(i).get
-        val from = fromJson[Float](to).get
-        (to must beEqualTo(i.toString)) and (from must beEqualTo(i))
-      }
+        basicMatcher(i, i.toString) }
     }
 
     case class Doubles() {
       def ok = forAll(arbitrary[Double]) { i =>
-        val to = toJson(i).get
-        val from = fromJson[Double](to).get
-        (to must beEqualTo(i.toString)) and (from must beEqualTo(i))
-      }
+        basicMatcher(i, i.toString) }
     }
 
     case class Dates() {
       def ok = forAll(arbitrary[DateTime]) { dt =>
-        val to = toJson(dt).get
-        val from = fromJson[DateTime](to).get
-        //currently Jackson-Joda serializer doesn't preserve timezone, which is what we want anyways
-        val expected = "\"%s\"".format(ISODateTimeFormat.dateTime().print(dt))
-        (to must beEqualTo(expected)) and (from must beEqualTo(dt))
-      }
+        basicMatcher(dt, "\"%s\"".format(ISODateTimeFormat.dateTime().print(dt))) }
     }
   }
 
   object Maps {
 
+    private def basicMapsMatcher[T : Manifest, U : Manifest](obj: Map[T, U], expectedJson: String) = {
+      val to = obj.toJson.get
+      val from = to.fromJson[Map[T, U]].get
+      val toConvert: Any = from.asInstanceOf[Any]
+      val converted = toConvert.convertValue[Map[T, U]].get
+
+      (to must beEqualTo(expectedJson)) and
+        (from.toSeq must containTheSameElementsAs(obj.toSeq)) and
+        (converted.toSeq must containTheSameElementsAs(from.toSeq))
+    }
+
+    private def listJson(lst: List[String]): String = {
+      val inner = (for {
+        li <- lst
+      } yield """"%s"""".format(li)).mkString(",")
+      s"[$inner]"
+    }
+
     case class StringToString() {
       def ok = forAll(genJsonString, genJsonString) { (k, v) =>
-        val to = toJson(Map(k -> v)).get
-        val from = fromJson[Map[String, String]](to).get
-
-        // string interpolation highlighting in triple-quoted strings doesn't render well in IntelliJ
-        (to must beEqualTo("""{"%s":"%s"}""".format(k, v))) and
-          (from must havePair(k -> v))
+        basicMapsMatcher(Map(k -> v), """{"%s":"%s"}""".format(k, v))
       }
     }
 
     case class StringToInt() {
       def ok = forAll(genJsonString, arbitrary[Int]) { (k, v) =>
-        val to = toJson(Map(k -> v)).get
-        val from = fromJson[Map[String, Int]](to).get
-
-        (to must beEqualTo("""{"%s":%d}""".format(k, v))) and
-          (from must havePair(k -> v))
+        basicMapsMatcher(Map(k -> v), """{"%s":%d}""".format(k, v))
       }
     }
 
     case class StringToListString() {
       def ok = forAll(genJsonString, nonEmptyListOf(genJsonString)) { (k, l) =>
-        val to = toJson(Map(k -> l)).get
-        val from = fromJson[Map[String, List[String]]](to).get
+        val to = Map(k -> l).toJson.get
+        val from = to.fromJson[Map[String, List[String]]].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[Map[String, List[String]]].get
 
-        val listJson = (for {
-          li <- l
-        } yield """"%s"""".format(li)).mkString(",")
-
-        (to must beEqualTo("""{"%s":[%s]}""".format(k, listJson))) and
+        (to must beEqualTo("""{"%s":%s}""".format(k, listJson(l)))) and
           (from.get(k) must beSome.like { case lst =>
             lst.toSeq must containTheSameElementsAs(l.toSeq)
-          })
+          }) and
+          (converted must beEqualTo(from))
       }
     }
 
     case class StringToListListString() {
       def ok = forAll(genJsonString, nonEmptyListOf(nonEmptyListOf(genJsonString))) { (k, l) =>
-        val to = toJson(Map(k -> l)).get
-        val from = fromJson[Map[String, List[List[String]]]](to).get
+        val to = Map(k -> l).toJson.get
+        val from = to.fromJson[Map[String, List[List[String]]]].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[Map[String, List[List[String]]]].get
 
-        from.get(k) must beSome.like { case lst =>
-          lst.toSeq must containTheSameElementsAs(l.toSeq)
-        }
+        val innerListString = (for {
+          innerList <- l
+        } yield listJson(innerList)).mkString(",")
+
+        (to must beEqualTo("""{"%s":[%s]}""".format(k, innerListString))) and
+          (from.get(k) must beSome.like { case lst =>
+            lst.toSeq must containTheSameElementsAs(l.toSeq)
+          }) and
+          (converted must beEqualTo(from))
       }
     }
 
     case class StringToListInt() {
       def ok = forAll(genJsonString, nonEmptyListOf(arbitrary[Int])) { (k, l) =>
-        val to = toJson(Map(k -> l)).get
-        val from = fromJson[Map[String, List[Int]]](to).get
+        val to = Map(k -> l).toJson.get
+        val from = to.fromJson[Map[String, List[Int]]].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[Map[String, List[Int]]].get
 
         val listJson = (for {
           li <- l
@@ -193,19 +199,23 @@ class JsonUtilSpecs
         (to must beEqualTo("""{"%s":[%s]}""".format(k, listJson))) and
           (from.get(k) must beSome.like { case lst =>
             lst.toSeq must containTheSameElementsAs(l.toSeq)
-          })
+          }) and
+          (converted must beEqualTo(from))
       }
     }
 
     case class StringToMapStringString() {
       def ok = forAll(genJsonString, genJsonString, genJsonString) { (k, k1, v1) =>
-        val to = toJson(Map(k -> Map(k1 -> v1))).get
-        val from = fromJson[Map[String, Map[String, String]]](to).get
+        val to = Map(k -> Map(k1 -> v1)).toJson.get
+        val from = to.fromJson[Map[String, Map[String, String]]].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[Map[String, Map[String, String]]].get
 
         (to must beEqualTo("""{"%s":{"%s":"%s"}}""".format(k, k1, v1))) and
           (from.get(k) must beSome.like { case m =>
             m must havePair(k1 -> v1)
-          })
+          }) and
+          (converted must beEqualTo(from))
       }
     }
   }
@@ -214,10 +224,13 @@ class JsonUtilSpecs
 
     case class ListOption() {
       def ok = forAll(nonEmptyListOf(genOption(genJsonString))) { l =>
-        val to = toJson(l).get
-        val from = fromJson[List[Option[String]]](to).get
+        val to = l.toJson.get
+        val from = to.fromJson[List[Option[String]]].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[List[Option[String]]].get
 
-        from.toSeq must containTheSameElementsAs(l.toSeq)
+        (from.toSeq must containTheSameElementsAs(l.toSeq)) and
+          (converted must beEqualTo(from))
       }
     }
 
@@ -226,40 +239,67 @@ class JsonUtilSpecs
   object CaseClasses {
     import com.paypal.cascade.json.tests.JsonUtilSpecs._
 
+    private def caseClassMatcher[T : Manifest](obj: T, expectedJson: String) = {
+      val to = obj.toJson.get
+      val from = to.fromJson[T].get
+      val toConvert: Any = from.asInstanceOf[Any]
+      val converted = toConvert.convertValue[T].get
+
+      (to must beEqualTo(expectedJson)) and
+        (from must beEqualTo(obj)) and
+        (converted must beEqualTo(from))
+    }
+
     case class OneMember() {
       def ok = forAll(genJsonString) { v =>
-        val to = toJson(OneMemberData(v)).get
-        val from = fromJson[OneMemberData](to).get
+        caseClassMatcher(OneMemberData(v), """{"value":"%s"}""".format(v))
+      }
+    }
 
-        (to must beEqualTo("""{"value":"%s"}""".format(v))) and
-          (from must beEqualTo(OneMemberData(v)))
+    case class OneMemberAny() {
+      def ok = forAll(genJsonString, arbitrary[Int], option(genJsonString)) { (v, i, mbS) =>
+        val stringData = OneMemberAnyData(v)//, """{"value":"%s"}""".format(v)) and
+        val intData = OneMemberAnyData(i)//, """{"value":%d}""".format(v))
+        val nestedData = OneMemberAnyData(stringData)
+        val optionalAnyData = OneMemberAnyData(OneMemberOptionalAnyData(mbS))
+
+        // these force a JSON conversion ping-pong, going to and from a JSON String,
+        // to make sure that we lose the original input data type along the way,
+        // and are forced to convert it out with `convertValue`
+        val sdJson = stringData.toJson.get.fromJson[OneMemberAnyData].get.value.convertValue[String].get
+        val idJson = intData.toJson.get.fromJson[OneMemberAnyData].get.value.convertValue[Int].get
+        val ndJson = nestedData.toJson.get.fromJson[OneMemberAnyData].get.value.convertValue[OneMemberAnyData].get
+        val ndJsonInner = ndJson.value.convertValue[String].get
+
+        // force a two-step conversion
+        val optionalAnyDataJson = optionalAnyData.toJson.get.fromJson[OneMemberAnyData].get
+        val optionalAnyDataInnerJson = optionalAnyDataJson.value.convertValue[OneMemberOptionalAnyData].get
+        val optionalAnyDataJsonValue = optionalAnyDataInnerJson.value.convertValue[Option[String]].get
+
+        (sdJson must beEqualTo(v)) and
+          (idJson must beEqualTo(i)) and
+          (ndJson must beEqualTo(stringData)) and
+          (ndJsonInner must beEqualTo(v)) and
+          (optionalAnyDataJsonValue must beEqualTo(mbS))
       }
     }
 
     case class TwoMemberMixedBasic() {
       def ok = forAll(genJsonString, arbitrary[Int]) { (s, i) =>
-        val to = toJson(TwoMemberMixedBasicData(s, i)).get
-        val from = fromJson[TwoMemberMixedBasicData](to).get
-
-        (to must beEqualTo("""{"one":"%s","two":%d}""".format(s, i))) and
-          (from must beEqualTo(TwoMemberMixedBasicData(s, i)))
+        caseClassMatcher(TwoMemberMixedBasicData(s, i), """{"one":"%s","two":%d}""".format(s, i))
       }
     }
 
     case class TwoMemberMixedComplex() {
       def ok = forAll(genJsonString, genJsonString, genJsonString, arbitrary[Int]) { (li1, li2, k, v) =>
-        val to = toJson(TwoMemberMixedComplexData(List(li1, li2), Map(k -> v))).get
-        val from = fromJson[TwoMemberMixedComplexData](to).get
-
-        (to must beEqualTo("""{"one":["%s","%s"],"two":{"%s":%d}}""".format(li1, li2, k, v))) and
-          (from must beEqualTo(TwoMemberMixedComplexData(List(li1, li2), Map(k -> v))))
+        caseClassMatcher(
+          TwoMemberMixedComplexData(List(li1, li2), Map(k -> v)),
+          """{"one":["%s","%s"],"two":{"%s":%d}}""".format(li1, li2, k, v))
       }
     }
 
     case class OptionalAnyValMember() {
       def ok = forAll(genJsonString, option(arbitrary[Int])) { (s, mbInt) =>
-        val to = toJson(OptionalAnyValData(s, mbInt)).get
-        val from = fromJson[OptionalAnyValData](to).get
 
         //None's should not be serialized
         val mbIntToJson: String = mbInt match {
@@ -267,15 +307,12 @@ class JsonUtilSpecs
           case Some(i) => ",\"mbTwo\":" + i
         }
 
-        (to must beEqualTo("""{"one":"%s"%s}""".format(s, mbIntToJson))) and
-          (from must beEqualTo(OptionalAnyValData(s, mbInt)))
+        caseClassMatcher(OptionalAnyValData(s, mbInt), """{"one":"%s"%s}""".format(s, mbIntToJson))
       }
     }
 
     case class OptionalAnyRefMember() {
       def ok = forAll(option(genJsonString), arbitrary[Int]) { (mbStr, i) =>
-        val to = toJson(OptionalAnyRefData(mbStr, i)).get
-        val from = fromJson[OptionalAnyRefData](to).get
 
         //None's should not be serialized
         val mbStrToJson = mbStr match {
@@ -283,18 +320,13 @@ class JsonUtilSpecs
           case Some(str) => """"mbOne":"%s",""".format(str)
         }
 
-        (to must beEqualTo("""{%s"two":%d}""".format(mbStrToJson, i))) and
-          (from must beEqualTo(OptionalAnyRefData(mbStr, i)))
+        caseClassMatcher(OptionalAnyRefData(mbStr, i), """{%s"two":%d}""".format(mbStrToJson, i))
       }
     }
 
     case class NestedClasses() {
       def ok = forAll(genJsonString) { s =>
-        val to = toJson(NestedOuter(NestedInner(s))).get
-        val from = fromJson[NestedOuter](to).get
-
-        (to must beEqualTo("""{"inner":{"one":"%s"}}""".format(s))) and
-          (from must beEqualTo(NestedOuter(NestedInner(s))))
+        caseClassMatcher(NestedOuter(NestedInner(s)), """{"inner":{"one":"%s"}}""".format(s))
       }
     }
 
@@ -304,19 +336,25 @@ class JsonUtilSpecs
 
     case class OptionalNested() {
       def ok = forAll(option(genNestedInner)) { mbN =>
-        val to = toJson(OptionalInner(mbN)).get
-        val from = fromJson[OptionalInner](to).get
+        val to = OptionalInner(mbN).toJson.get
+        val from = to.fromJson[OptionalInner].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[OptionalInner].get
 
-        from must beEqualTo(OptionalInner(mbN))
+        (from must beEqualTo(OptionalInner(mbN))) and
+          (converted must beEqualTo(from))
       }
     }
 
     case class ListOptionMember() {
       def ok = forAll(nonEmptyListOf(genOption(genJsonString))) { l =>
-        val to = toJson(ListOptionData(l)).get
-        val from = fromJson[ListOptionData](to).get
+        val to = ListOptionData(l).toJson.get
+        val from = to.fromJson[ListOptionData].get
+        val toConvert: Any = from.asInstanceOf[Any]
+        val converted = toConvert.convertValue[ListOptionData].get
 
-        from.l.toSeq must containTheSameElementsAs(l.toSeq)
+        (from.l.toSeq must containTheSameElementsAs(l.toSeq)) and
+          (converted must beEqualTo(from))
       }
     }
   }
@@ -340,19 +378,19 @@ class JsonUtilSpecs
         val unfinishedInnerMap2 = """{"%s":{"%s":"%s"}""".format(k, v, v)
         val unfinishedInnerMap3 = """{"%s":"%s":"%s"}}""".format(k, v, v)
 
-        (fromJson[Map[String, String]](unquotedKey) must beFailedTry) and
-          (fromJson[Map[String, String]](partiallyQuotedKey1) must beFailedTry) and
-          (fromJson[Map[String, String]](partiallyQuotedKey2) must beFailedTry) and
-          (fromJson[Map[String, String]](unquotedStringValue) must beFailedTry) and
-          (fromJson[Map[String, String]](partiallyQuotedStringValue1) must beFailedTry) and
-          (fromJson[Map[String, String]](partiallyQuotedStringValue2) must beFailedTry) and
-          (fromJson[Map[String, String]](noOpeningBrace) must beFailedTry) and
-          (fromJson[Map[String, String]](noClosingBrace) must beFailedTry) and
-          (fromJson[Map[String, List[String]]](unfinishedInnerList1) must beFailedTry) and
-          (fromJson[Map[String, List[String]]](unfinishedInnerList2) must beFailedTry) and
-          (fromJson[Map[String, Map[String, String]]](unfinishedInnerMap1) must beFailedTry) and
-          (fromJson[Map[String, Map[String, String]]](unfinishedInnerMap2) must beFailedTry) and
-          (fromJson[Map[String, Map[String, String]]](unfinishedInnerMap3) must beFailedTry)
+        (unquotedKey.fromJson[Map[String, String]] must beFailedTry) and
+          (partiallyQuotedKey1.fromJson[Map[String, String]] must beFailedTry) and
+          (partiallyQuotedKey2.fromJson[Map[String, String]] must beFailedTry) and
+          (unquotedStringValue.fromJson[Map[String, String]] must beFailedTry) and
+          (partiallyQuotedStringValue1.fromJson[Map[String, String]] must beFailedTry) and
+          (partiallyQuotedStringValue2.fromJson[Map[String, String]] must beFailedTry) and
+          (noOpeningBrace.fromJson[Map[String, String]] must beFailedTry) and
+          (noClosingBrace.fromJson[Map[String, String]] must beFailedTry) and
+          (unfinishedInnerList1.fromJson[Map[String, List[String]]] must beFailedTry) and
+          (unfinishedInnerList2.fromJson[Map[String, List[String]]] must beFailedTry) and
+          (unfinishedInnerMap1.fromJson[Map[String, Map[String, String]]] must beFailedTry) and
+          (unfinishedInnerMap2.fromJson[Map[String, Map[String, String]]] must beFailedTry) and
+          (unfinishedInnerMap3.fromJson[Map[String, Map[String, String]]] must beFailedTry)
       }
     }
 
@@ -365,20 +403,80 @@ class JsonUtilSpecs
         forAll(
           genJsonString,
           genJsonString.suchThat(s => Try {Integer.decode(s)}.isFailure)) { (k, v) =>
-          val to = toJson(Map(k -> v)).get
+          val to = Map(k -> v).toJson.get
 
           // Note: if instead `v` were an Int, and `fromJson[Map[String, String]]` were used,
           // Jackson would attempt to convert the value to a String, which would be a Success
-          val from = fromJson[Map[String, Int]](to)
+          val from = to.fromJson[Map[String, Int]]
 
-          from must beFailedTry
+          val toConvert: Any = from.asInstanceOf[Any]
+          val converted = toConvert.convertValue[Map[String, Int]]
+
+          (from must beFailedTry) and (converted must beFailedTry)
+        }
+      }
+    }
+
+    // There are probably more variations that could be tested here.
+    case class BadValueConversions() {
+      def fails = {
+        forAll(
+          genJsonString,
+          nonEmptyListOf(arbitrary[Int]),
+          nonEmptyListOf(genJsonString.suchThat(s => Try {Integer.decode(s)}.isFailure))) { (s, il, sl) =>
+          val omd = OneMemberData(s)
+
+          /*
+           We can't test List[Int] => List[String] because of
+           the aforementioned quirk regarding converting Int to String.
+
+           What happens instead is that the List[Int] gets successfully converted
+           to a List[String] because Jackson knows that Ints have a String
+           representation. This is a known quirk, and we test for it elsewhere.
+           */
+
+          (omd.asInstanceOf[Any].convertValue[String] must beFailedTry) and
+            (omd.asInstanceOf[Any].convertValue[List[Int]] must beFailedTry) and
+            (omd.asInstanceOf[Any].convertValue[List[String]] must beFailedTry) and
+            (s.asInstanceOf[Any].convertValue[OneMemberData] must beFailedTry) and
+            (s.asInstanceOf[Any].convertValue[List[Int]] must beFailedTry) and
+            (s.asInstanceOf[Any].convertValue[List[String]] must beFailedTry) and
+            (il.asInstanceOf[Any].convertValue[OneMemberData] must beFailedTry) and
+            (il.asInstanceOf[Any].convertValue[String] must beFailedTry) and
+            (sl.asInstanceOf[Any].convertValue[OneMemberData] must beFailedTry) and
+            (sl.asInstanceOf[Any].convertValue[String] must beFailedTry) and
+            (sl.asInstanceOf[Any].convertValue[List[Int]] must beFailedTry)
+        }
+      }
+    }
+
+    case class AnyToNullConversion() {
+      def ok = {
+        forAll(option(genJsonString)) { mbS =>
+          val data = OneMemberAnyData(mbS)
+          val json = data.toJson.get.fromJson[OneMemberAnyData].get
+          val value = json.value.convertValue[Option[String]].get
+
+          /*
+           This was discovered during testing: if you serialize a None, which with the settings in
+           this module are simply not serialized, and then deserialize it to an Any, Jackson doesn't know
+           what to do and will deserialize a null instead.
+
+           If that behavior is fixed in the future, this test will begin failing. :)
+           */
+          mbS match {
+            case Some(s) => value must beSome.like { case v: String =>
+              v must beEqualTo(s)
+            }
+            case None => value must beNull
+          }
         }
       }
     }
 
     case class MissingAnyVal() {
       def ok = forAll(genJsonString) { k =>
-        val from = fromJson[TwoMemberMixedBasicData]("""{"one":"%s"}""".format(k)).get
+        val from = """{"one":"%s"}""".format(k).fromJson[TwoMemberMixedBasicData].get
 
         // this is a known Jackson quirk: missing AnyVals in case classes get instantiated at default values, e.g. 0
         from must beEqualTo(TwoMemberMixedBasicData(k, 0))
@@ -387,29 +485,19 @@ class JsonUtilSpecs
 
     case class MissingAnyRef() {
       def ok = forAll(arbitrary[Int]) { k =>
-        val from = fromJson[TwoMemberMixedBasicData]("""{"two":%d}""".format(k)).get
+        val from = """{"two":%d}""".format(k).fromJson[TwoMemberMixedBasicData].get
 
         // this is another known Jackson quirk: missing AnyRefs in case classes come in as `null`
         from must beEqualTo(TwoMemberMixedBasicData(null, k))
       }
-    }
-
-  }
-
-  case class ImplicitClasses() {
-    import com.paypal.cascade.json._
-    def success = forAll(genJsonString) { str =>
-      val to = str.toJson.get
-      val from = to.fromJson[String].get
-      // for whatever reason, the compiler balks on string interpolation here
-      // e.g. "\"$str\"" is not recognized as a valid string
-      (to must beEqualTo("\"%s\"".format(str))) and (from must beEqualTo(str))
     }
   }
 }
 
 object JsonUtilSpecs {
   case class OneMemberData(value: String)
+  case class OneMemberAnyData(value: Any)
+  case class OneMemberOptionalAnyData(value: Option[Any])
   case class TwoMemberMixedBasicData(one: String, two: Int)
   case class TwoMemberMixedComplexData(one: List[String], two: Map[String, Int])
   case class OptionalAnyValData(one: String, mbTwo: Option[Int])